+++ tokenize.hpp
@@ -54,7 +54,7 @@ public:
 
 	bool is_equal(std::string &&op) const;
 	bool is_equal(const std::string &op) const;
-	int get_number() const;
+	int64_t get_number() const;
 	bool is_typename() const;
 
 	/* 静的メンバ関数 (public) */
